Edit potential functions
Modified the potential functions so that it considers x,y positions rather than just r to more accurately get the evolution of the angles and radial distances to the bodies. I may still need to modify the U_cf functions because I am not sure if the masses are evolving correctly quite yet.